braille_converter/arduino_integration/requirements.txt
# Arduino Integration Requirements

# Serial communication
pyserial>=3.5

